requirements.txt
# Core dependencies
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Search and web utilities
tavily-python>=0.3.0

# Display and formatting
rich>=13.0.0
ipython>=8.0.0

# Type hints
typing-extensions>=4.0.0
